+++ src/main/java/com/digitalsanctuary/spring/user/persistence/model/User.java
@@ -111,7 +111,7 @@ public class User implements Serializable {
 	/** The roles - stored as Set to avoid Hibernate immutable collection issues */
 	@ToString.Exclude
 	@EqualsAndHashCode.Exclude
-	@ManyToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE}, fetch = FetchType.EAGER)
+	@ManyToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE}, fetch = FetchType.LAZY)
 	@JoinTable(name = "users_roles", joinColumns = @JoinColumn(name = "user_id", referencedColumnName = "id"),
 			inverseJoinColumns = @JoinColumn(name = "role_id", referencedColumnName = "id"))
 	private Set<Role> roles = new HashSet<>();